requirements.txt
django
djangorestframework
scikit-learn
imbalanced-learn
pandas
numpy
nltk
matplotlib
#pickle
wordcloud
gunicorn
uvicorn
psycopg2-binary
dj-database-url
whitenoise